sensor_combined.msg: use uint32 for integral_dt
There is no reason to make this 64 bit. The same should be done in
the sensor raw messages, together with further cleanup.

+++ src/modules/ekf2_replay/ekf2_replay_main.cpp
@@ -378,7 +378,7 @@ void Ekf2Replay::setEstimatorInput(uint8_t *data, uint8_t type)
 		parseMessage(data, dest_ptr, type);
 		_sensors.timestamp = replay_part1.time_ref;
 		_sensors.gyro_integral_dt = replay_part1.gyro_integral_dt;
-		_sensors.accelerometer_integral_dt = replay_part1.accelerometer_integral_dt;
+		_sensors.accelerometer_integral_dt = (uint32_t)replay_part1.accelerometer_integral_dt;
 		_sensors.magnetometer_timestamp = replay_part1.magnetometer_timestamp;
 		_sensors.baro_timestamp = replay_part1.baro_timestamp;
 		_sensors.gyro_integral_rad[0] = replay_part1.gyro_integral_x_rad;
